documentation error: reference to a file that does not exist
The Getting Started Guide for developers explains that in order to
run e2e tests while focusing on only a single test file, the user
can set the FUNC_TEST_ARGS variable and indicate a test file to focus on.
The file that it references does not exist, but the execution does not
indicate that and basically performs a no-op.
Fix the file reference so a file with a similar name that does exist

Signed-off-by: Nir Dothan <ndothan@redhat.com>