tests/network: simplify [test_id:1775]should not request a tun device
The fact that virt-launcher of a network-less VM does not require any
tun device is a minor implementation detail. It is not the business of
an end-to-end test to worry about it.

Using libpod.GetPodByVirtualMachineInstance to expect virt-launcher
structure is a code-smell, because a good end-to-end test of VMIs should
not care about pods, only about KubeVirt API and guest behavior. Pods
are KubeVirt implementation detail.

This test is simplified to ensure we can start a VMI with no network
insterface.

Signed-off-by: Dan Kenigsberg <danken@redhat.com>