Sound tests: Remove test
Remove that doesn't give us new value. The test
we left covers the same path + defaulting of model.

The we remove just test that supported model can
be explicitly supplied but this is already covered by
unit test. Other than that it checks xml schema which
is also covered by unit test in the converter.

Signed-off-by: Luboslav Pivarc <lpivarc@redhat.com>