tests: usbredir: avoid flakyness
For this test to work, we need 4 ongoing connections to usbredir
websocket for the 5th to fail.

In the CI, it was shown it is possible for some failure to happen in
any of the first 4 connections, resulting in a non failure of the 5th.

The test is improved making sure that we have 4 connections before
testing the 5th.

Signed-off-by: Victor Toso <victortoso@redhat.com>