Allow modifications to VM spec during VM snapshot, except disks/volumes
This change allows modifications to the VM spec
during a VM snapshot, except for the disks/volumes
field. This is to prevent users from changing the
disks/volumes field during a snapshot, which
could cause issues with the snapshot process.

Instead of rejecting all changes to the VM spec.
We allow changes that don't modify disks/volumes.
If we don't allow this then mutating webhooks can
prevent the snapshot from getting a lock on the VM,
because it can't apply the snapshot finalizer.

Example is an old version of the VM spec doesn't
include the architecture, and we attempt to take
a snapshot with a version of KubeVirt that requires
architecture. The mutating webhook will attempt to
add the architecture field to the VM spec, but it
is rejected because the status indicates that the
VM is attempting a snapshot, and thus rejecting the
change. Now we can't get a lock due to the missing
finalizer, and the snapshot will never complete.

Signed-off-by: Alexander Wels <awels@redhat.com>