virt-handler, net: Load the 'macvtap' kmod implicitly
In case the `managedMacvtap` `domainAttachmentType` is used, a `macvtap`
link is created in the `virt-launcher` pod. However, on nodes which have
security measures enabled (e.g. SELinux), the loading of the `macvtap`
kmod may fail.
The kmod is usually loaded per need implicitly.

Therefore, Kubevirt needs to assure the kmod is loaded in advance to
avoid failure of the potential `macvtap` kmod loading.

The method chosen to load the kmod in advance is through the creation of
a temp `macvtap` link. This avoids the need to depend on `modprobe` and
the location of the kmod/s on the OS distribution.

The temp link is created on `virt-handler` initial start.

Signed-off-by: Edward Haas <edwardh@redhat.com>