API, VMI interface status: Add PodInterfaceName
Add the `PodInterfaceName` fields to the interface status struct.
The field exposes the pod interface name on the VMI status interfaces
in order for other entities to consume it per need.

It may also be useful for troubleshooting the pod netns issues,
identifying which pod interface is connected to which network.

Signed-off-by: Orel Misan <omisan@redhat.com>