Change NonDefaultNetworkStatusIndexedByIfaceName's signature
Currently, the function takes a pod as an input.
Change the function's signature to take a slice of NetworkStatus
objects.
Doing so:
1. Simplifies the implementation by working on a processed input.
2. Simplifies the unit tests by providing structs as inputs
instead of raw JSON strings.

The test with ordinal naming scheme was dropped
because naming scheme does not affect the logic.

Also rename it to better reflect its functionality.

Signed-off-by: Orel Misan <omisan@redhat.com>