instancetype: Extract vm-controller sync logic
This commit starts the process of extracting instance type and
preference logic from the VirtualMachine controller, starting with the
core sync loop logic.

The existing synchronizer interface is used as the foundation of this
implementation. However for the time being the actual call to this
implementation is left where the original logic resided, well before the
network synchronizer implementation is called. A TODO is left in place
to eventually move the instancetype synchronizer alongside this.

Signed-off-by: Lee Yarwood <lyarwood@redhat.com>