Only create VMExport token secret if there are volumes to export
If there are no volumes to export, and the tokenSecretRef was
not set, the VMExport controller would try to create a secret
endlessly. It would succeed the first time, and each subsequent
time it would fail because the secret already exists.

This was causing noise in the API server about creating
a secret that already exists. This patch adds a check
to only create the secret if there are volumes to export.

This should resolve it self quickly if there are available
volumes.

Signed-off-by: Alexander Wels <awels@redhat.com>