virt-handler, net setup: Use naming scheme from status
Currently, the network setup logic in the virt-handler
calculates the pod interface naming scheme locally based
on information that discovered by nmstate.

Use the information from VMI.Status.Interfaces instead.
Doing so:
1. Centralizes the calculation in one place (virt-controller).
2. Makes the local calculation redundant.
3. Allows for an easier change in the future.
4. Enables using a custom primary interface name (other than "eth0").

Don't start performing the network setup in case the naming
information is missing.
It could happen on a NIC hotplug scenario, and will be resolved in
a following reconcile.

Signed-off-by: Orel Misan <omisan@redhat.com>