Allow vmexport to have names longer than 63 characters
Nothing was preventing the name to be longer than 63
characters, but it would fail to create the export
service and associated labels because it was using the
vmexport name as part of the label. And labels cannot
be longer than 63 characters.

Also ensured the container name is also not longer
than 63 characters. This was generating a warning
in log that it was not recommended.

Signed-off-by: Alexander Wels <awels@redhat.com>