Allow dots in VM volume names
Volume names are allowed to contain dots, however
the VMExport was using the volume name as part of
naming convention in the pod spec. The name in the
pod spec does not allow dots.

This patch replaces all dots with dashes in the
volume name so the pod can be created successfully.

Signed-off-by: Alexander Wels <awels@redhat.com>
(cherry picked from commit bba31d488570dd0b560a66d9bc465532bc4ca7a5)