fix(virtctl,create): Fix boot order of volumes
Ensure the documented boot order and the actual boot order of volumes
added to a VM match. This change is also a preparation for deprecating
some of the existing flags and aliasing them to --volume-import.

Once the flags are dropped only Imported volumes will remain and
Containerdisks and PVC will always come first in the order.

Signed-off-by: Felix Matouschek <fmatouschek@redhat.com>