Fix virtctl imageupload ds handling
The `handleDataSource` function checks the existence
of the ds and will update or create it based on the response.

Currently, the check is based on the ds returned by the client
```
ds != nil
```
This is wrong, since the client always returns an object,
empty in case of error.
Since the not found error is tolerated in the previous statement
this results in a wrong handling of the creation.

Signed-off-by: fossedihelm <ffossemo@redhat.com>