Manage Creation of UI configuration ConfigMaps by HCO (#3136)
Currently, when kubevirt-hco is installed, the console-plugin
configuration configmaps are created only when a cluster-admin opens the
Virtualization tab for the first time. At that point two ConfigMaps are
created in the installation namespace:
- kubevirt-user-settings
- kubevirt-ui-features

And also their RBACs to be reachable by regular users are created.
Without login by an admin, the config maps are missing, and regular
users cannot use the UI at that point.

This PR adds logic to create those configmaps and associated RBACs, if
they are not already exist.
Also, remove the ConfigMaps from their previous location at the default
namespace, used in v1.12.0 and below.

Signed-off-by: Oren Cohen <ocohen@redhat.com>