virt-config, feature-gates: Graduate the network-binding-plugins to Beta
The network binding plugins feature has been introduce back in v1.1.0,
over a year ago.
Since, it has seen wide adoption in the community with multiple plugin
vendors that introduced various network bindings through it.

It seems that the feature has reached a level of maturity, allowing it
to graduate to Beta and be targeted for GA in the next release.

The network binding plugins feature has a Feature Gate which is now
marked as Beta in the next Kubevirt release (v1.4).

Signed-off-by: Edward Haas <edwardh@redhat.com>