feat(virtctl,create): Add support for access credentials
Add the --access-cred flag to virtctl create vm which allows to specify
access credentials (ssh public keys or passwords), which are injected
into a VM by KubeVirt. The secret used to store public keys or
passwords must already exist.

Depending on the chosen options a cloud-init configuration is generated
to allow the qemu guest agent to manage ssh keys or to allow the chosen
propagation method (noCloud or configDrive) to work.

Signed-off-by: Felix Matouschek <fmatouschek@redhat.com>