Retry calculating the md5sum of the export disk
A verification pod is created to calculate the
md5sum of the export disk. This pod uses fsGroup
to give the correct user access to the disk.img.

Sometimes the md5sum is executed before the chmod
has had a chance to complete. This change adds a
retry so that the md5sum can hopefully be calculated
after the chmod completes.

Signed-off-by: Alexander Wels <awels@redhat.com>