cleanup(virtctl,create): Support mixed cases in params
Convert the params passed to RunStrategyFlag, SysprepVolumeFlag,
InstancetypeFlag and PreferenceFlag to lowercase before comparing them
to known values.

Conflicts:
- pkg/virtctl/create/vm/vm_test.go

Note:
The conflict is due to a test not being updated for the lower-cased
values. This is a mistake in the original PR, which was masked by the
fix being included in a further commit.

Signed-off-by: Felix Matouschek <fmatouschek@redhat.com>