io_utils: move ExecuteCommandInVirtHandlerPod to libnode
ExecuteCommandInVirtHandlerPod has nothing to do with io. On the other
hand, we have a nearly-identical function
libnode.ExecuteCommandOnNodeThroughVirtHandler. Let us move the two
functions next to each other and eliminate the catch-all junkyard file
that was io_utils.

ExecuteCommandInVirtHandlerPod and
ExecuteCommandOnNodeThroughVirtHandler are left confusingly similar. I
hope we can improve their situation in the future.

Signed-off-by: Dan Kenigsberg <danken@redhat.com>