virt-controller, virt-config: Wire up referencePolicy
This commit wires up the newly introduced reference policies within the
VM controller and virt-config.

The default policy will be reference with no changes to the default
behaviour of instance types and preferences.

An additional policy of expand is also wired up. When explicitly enabled
this will expand VirtualMachines without revisionNames in the same way
as the expand-spec API and VM subresource API.

TODOs are left in place for an upcoming refactor of the instance type
package that will impact how callers such as the VM controller make use
of it.

Signed-off-by: Lee Yarwood <lyarwood@redhat.com>