Fix `check-unassigned-tests` script
`check-unassigned-tests` checks that there are no tests that are
not assigned to any sig.
To do that it ensures that every tests has `[sig-<sig_name>]` pattern
in the test descriptor.

This is not enough, since our `automation/test.sh` script is not
using file name based filtering anymore, but label filtering.

This means that, in case a test has the `[sig-<sig_name>]` string in
its name, but not the label, the check-unassigned-tests.sh passes but the
test is actually not executed.

Let's introduce the double check: the legacy file name based check and the
new label check.

Also, the `sig-compute-migrations` was not covered by this check, so introduced
it too.

Signed-off-by: fossedihelm <ffossemo@redhat.com>