net, deviceinfo: Extract network-status mapping logic
Extract the logic responsible for mapping network status by pod interface name
to `pkg/network/status`.

According to the "Kubernetes Network Custom Resource Definition
De-facto Standard" [1], the `Interface` field of the network status struct
is optional and could be empty.

This means that we could have a key equal to an empty string mapping
to the last `NetworkStatus` struct that its interface field was empty.

Since there are currently only read operations from the map returned
by `NetworkStatusesByPodIfaceName` this is acceptable.

[1] https://github.com/k8snetworkplumbingwg/multi-net-spec/tree/master/v1.3

Signed-off-by: Orel Misan <omisan@redhat.com>