net, namescheme: Introduce CreateFromNetworkStatuses
Currently, `CreateNetworkNameSchemeByPodNetworkStatus`
is used to create the naming scheme based on the pod's network-status.

Introduce `CreateFromNetworkStatuses` as a replacement that takes
in a slice of `NetworkStatus` objects.
Doing so enables us to remove the mapping of
`NetworkStatus` objects by pod interface name in
VMI controller's `updateMultusAnnotation` method.

`CreateNetworkNameSchemeByPodNetworkStatus` will be removed in the
following commit.

Signed-off-by: Orel Misan <omisan@redhat.com>