virt-controller, vm: sync Firmware.UUID into vm.spec
As described in https://github.com/kubevirt/kubevirt/issues/13156
our code for Firmware.UUID is buggy, generating non universally-unique
IDs.

This commit keeps the buggy logic so as not to modify the ID of existing
VMs, but it makes sure that the Firmware.UUID is persisted in the VM
spec. In two or three releases we can be sure that all existing VMs have
their Firmware.UUID persisted and can fix the logic to the correct (and
simpler)

    vm.Spec.Template.Spec.Domain.Firmware.UUID = types.UID(uuid.New())

The new syncFirmwareUUID() is called much more often than
setupStableFirmwareUUID() so logging each action would not be tenable.

Signed-off-by: Dan Kenigsberg <danken@redhat.com>