Centralize the arch converter
Currently, the arch converter is used
to differentiate the behavior between
the arch.
To do so, it should be initialized every time
you want to use it (`newArchConverter(arch)`).

The converterContext has the architecture string
as field.

We can expand the latter to the archConverter, so
that we can use it directly.

Ideally, the archConverter should be the centralized
place where all the different strategies are contained,
without the needs to specify `isARM` or `isAMD` spread
in the codebase.
I'd expect to see something like `is<feature>Enabled()`
where the different archconverter handle it.

Signed-off-by: fossedihelm <ffossemo@redhat.com>