fix: Properly restore kv.spec.uninstallStrategy
Currently serial tests rely on the global cleanup.
Actually the global cleanup just ensure the kv.spec.config
is restored, not the entire spec.
Test id [test_id:3683] changes the uninstallStrategy to
`BlockUninstallIfWorkloadsExist` but does not restore the
previous setting.
This can lead to failure to the subsequent tests, for example
`[test_id:3146]should be able to delete and re-create kubevirt install`,
in which uninstall strategy is evaluated.

Signed-off-by: fossedihelm <ffossemo@redhat.com>