vmi-ctrl: Decouple NIC hotplug / hotunplug logic
Currently, the logic responsible for NIC hotplug / hotunplug
resides in the VMI controller and is unit tested in its context.

Extract this logic to the network annotation generator - which
decouples it from the VMI controller.

Another advantage of this move is the fact that the
virt-launcher pod is only patched once per sync cycle
with all the network-related annotations instead of twice
in the worst case.

Prior to this change, the NIC hotplug / unplug logic was
called after SIG-storage's volume hotplug logic.
Following this change, the network code is called first as
part of an existing network logic.
The storage code should not be affected by this change, since:
1. The network logic only affects the virt-launcher pod's annotations.
2. There is already a patch operation of the virt-launcher
pod's annotations prior to the volume hotplug sequence.

Signed-off-by: Orel Misan <omisan@redhat.com>