Allow VMExport to start if completed VMI exists
It is possible that a VMI has completed or failed
and one wants to export the VM disk. This change
allows the VMExport to become ready if either of
those conditions are met.

The default behavior if a VM is stopped normally
is to delete the VMI, but in case this doesn't
happen, the VMExport should still be allowed to
start.

Signed-off-by: Alexander Wels <awels@redhat.com>