Graduate the DynamicPodInterfaceNaming FG
The DynamicPodInterfaceNaming feature-gate was introduced
in v1.4 by PR #13078.

The feature works well for at least two projects.
No negative feedbacks were received from the community
on the implementation other than the fact that is does
not support migration between two different primary interface
names (in this case the migration will fail).

This scenario is not supposed to be supported for the
foreseeable future.

Signed-off-by: Orel Misan <omisan@redhat.com>