fix: Proactively trigger handler final cleanup when a migration fails
Currently, when a migration fails we signal the target pod to cleanup,
and then we rely on a domain change event from the informer to unmount all
the volumes (containerdisks, etc...).
If the domain was never created (just for example: using `FuncTestForceLauncherMigrationFailureAnnotation`),
the virt-handler is informed of the stale client connection only once the watchdog timeout expires (30 seconds).

This delay can cause some tests to fail, for example "old finalized migrations should get garbage collected".

This can be avoided by re-enqueueing the vmi object when the handler is acknowledged
that the migration has failed.

Signed-off-by: fossedihelm <ffossemo@redhat.com>