vmsnapshot: update error on content and snapshot if freeze fails
Up until now if the freeze failed in the online
snapshot it was not reflected on the snapshot object
and eventually the deadline would exceed and the user
would not understand what happened.
Now in case freeze fails it will be marked as an error
on the vmsnapshotcontent which is reflected also on the vmsnapshot.
Remove logic of skiping volumesnapshots in case of error since
error should be able to change and then continue with the snapshot
with previous code if an error occured you will keep skipping
and putting an error on the content without exiting the error state.

Signed-off-by: Shelly Kagan <skagan@redhat.com>