Mv calculateInterfacesAndNetworksForMultusAnnotationUpdate
The said function is only used by the net annotations-gen.
Move it closer to its usage and test it as part of its caller.

The removed unit tests are covered by the network annotations
generator tests:

calculate if changes are required
- when vmi interfaces have an extra interface which requires hotplug
Is covered by [1].

- when a vmi interface has state set to `absent`, requiring hotunplug
This scenario is blocked by the VM controller [2], and was removed
because it does not reflect the expected input.

- when vmi interfaces have an interface to hotplug and one to hot-unplug,
 given hashed names
Is covered by [3].

[1] https://github.com/kubevirt/kubevirt/blob/ba9ddf19293c30ab88864060031c699756a7ac0b/pkg/network/pod/annotations/generator_test.go#L613
[2] https://github.com/kubevirt/kubevirt/blob/ba9ddf19293c30ab88864060031c699756a7ac0b/pkg/virt-controller/network/hotplug.go#L34
[3] https://github.com/kubevirt/kubevirt/blob/ba9ddf19293c30ab88864060031c699756a7ac0b/pkg/network/pod/annotations/generator_test.go#L640

Signed-off-by: Orel Misan <omisan@redhat.com>