fg: graduation of volume migration
The VolumeUpdateStrategy and VolumeMigration feature gates were
introduced in v1.3.0 by #11533.

Both feature gates have been present for 2 releases and since then
multiple issues have been reported and fixed.
With the graduation of the two feature gates, we want to indicate to
KubeVirt users that the feature will remain and that the API may be
considered stable for use.

Signed-off-by: Alice Frosi <afrosi@redhat.com>