Fix volume migration tests against ceph rbd
Some of the tests where expecting filesystem RWO
volumes, but when using the WithStorage functions
the tests will use storage profile to fill out
missing information. In some cases if the storage
identified was ceph-rbd it would create RWX block
volumes instead of RWO filesystem volumes.

This fixes those cases to explicitly create RWO
filesystem volumes.

In one case a PVC is created directly when migrating
from block to filesystem. And the filesystem volume
did not calculate fsOverhead and thus was not large
enough to complete the migration. This increases the
size of the target volume to ensure there is enough
space for the migration to complete.

Signed-off-by: Alexander Wels <awels@redhat.com>