hotplug volume: Boot from hotpluggable disk
Currently, virt-handler does not wait for the hotplug to be mounted
before attempting sync the libvirt domain. This is problematic if
we want to support booting from usb disk or ejectable cdrom for example

Fix is to simply wait for the hotlug pod volume to be mounted. We already do
something similar with containerdisks. Have to wait for them to be pulled

Signed-off-by: Michael Henriksen <mhenriks@redhat.com>