VMRestore: move vmsnapshot checks to main function
Having a valid vmsnapshot for the restore is a
basic operation and should not be in the context
of another.
On top of that, added a check that was only in the vmrestore
admitter of making sure the restore doesnt override an existing
target. This is a racy check and should be done in the reconciler
all the time and not in the webhook.

Signed-off-by: Shelly Kagan <skagan@redhat.com>