vm: validate hotplugged volume for VM spec update
In case of VM rollout strategy isn't live-update, we still need to take
into account that the volume set could be changed because of hotplugged
volumes.

Hotplugged volumes have been added before the live-updatable fields and
they follow a different path then cpu and memory hotplug. The volumes
addition and removal goes through the api subresource and they are
imperatively added to the spec.

Signed-off-by: Alice Frosi <afrosi@redhat.com>