vm: test the hotplug volumes and restart condition
The first unit test verifies that the restart condition isn't set
when the vm rollout strategy is stage, and the vm spec are modified
because by an imperative volume hotplug request.

This unit test reproduce the case reported in:
  https://github.com/kubevirt/kubevirt/issues/13222

The second unit test verifies that the restart condition is set when
the vm rollout strategy is stage, and the vm spec are modified because
a declarative hotplug.

Signed-off-by: Alice Frosi <afrosi@redhat.com>