access-credentials: test: Fix race in unit test.
One unintuitive but intentional behavior in Ginkgo is that
all "DeferCleanup()" blocks are executed after all "AfterEach()" blocks.
This causes a race in this test case, because "AfterEach()" removes a file
that is being watched by a goroutine inside HandleQemuAgentAccessCredentials().
If the goroutine is faster, it will call LookupDomainByName() second
time and panic.

Using "defer" instead of "DeferCleanup()" should fix the race.

Signed-off-by: Andrej Krejcir <akrejcir@redhat.com>