virt-config, fg: Graduate the Network Binding Plugin feature to GA
The network binding plugins feature has been introduce back in v1.1.0,
over a year ago.
Since, it has seen wide adoption in the community with multiple plugin
vendors that introduced various network bindings through it.

The feature has reached a level of maturity, allowing it to be declared
GA for the coming release.

The network binding plugins feature has a Feature Gate which is now
marked as GA in the next Kubevirt release (v1.5).

Production and e2e testing code no longer depend on it.

Signed-off-by: Edward Haas <edwardh@redhat.com>