refactor(CNV-51887): move libdv out of functional tests
The libdv package contains NewDataVolume
function options that helps us to construct
a new v1.VirtualMachine.

Currently libdv used only in functional tests
and not importable by other projects, because
tests likely tend to be changed or removed.

Move libdv package to pkg folder that is likely
not to be changed like libvmi package.

Signed-off-by: Ben Oukhanov <boukhanov@redhat.com>