hooks: refactor collectSideCarSockets
This commit refactors collectSideCarSockets to poll the hook sidecar
sockets every second with a specified timeout, replacing the previous
for and sleep approach.

It concurrently processes the sockets in the sockets directory,
continuing until either all required sockets (based on the number of
requested hook sidecards) have been processed, a failure occurs, or the
timeout is reached.

Signed-off-by: Adi Aloni <adialoni101@gmail.com>