net, setup: Set const name for tap associated with VM's primary iface
Currently, tap devices generated by KubeVirt have a dynamic
name that is derived from their associated pod interface name.

Following PR #13018, it is possible to have a dynamic
name for the pod's primary interface (other than `eth0`).

At the moment, it is not possible to migrate to a target
pod that has a primary interface name different from the source,
as the tap name is set in the domain XML.

In order to provide an option for a futuristic ability to
accommodate for this kind of migration - set the tap name
associated with the VM's primary interface to `tap0`.

Doing so will allow migrating the domain XML without
modifying it either at the source or at the target.

An additional API change would be required in order to support
such a migration, but it could be differed to when the need
arises.

Signed-off-by: Orel Misan <omisan@redhat.com>