tests, net, vmi_networking: Assert vendor via pci
set a PCI address to interface
and assert device vendor according to PCI adress
this enhacement is needed becaue we currently find the
vendor file according to its interface name, and it was hardcoded
to eth0. As we are adding another interface, the ordinal of eth
ifaces is undeterministic and a different method is required in
order to find the vendor file per virtual device. Hence set an explicit
PCI address per Iface and then find their vendor file in the guest
FS according to PCI address

Signed-off-by: Nir Dothan <ndothan@redhat.com>