e2e guestlog: cleanup
e2e tests are expensive (in CI budget and human time),
and fragile (raceful by design).
They should be limited to actual e2e test,
which test only KubeVirt APIs and features,
not implementation details.
The DescribeTable `should successfully start with LogSerialConsole`
goes against this principle and that logic is already
covered by unit test.
For this reason, it can be dropped.
Also, a sort imports is performed

Signed-off-by: fossedihelm <ffossemo@redhat.com>