cleanup(tests,virtctl): Make virtctl helpers unexported
Move the helpers to run virtctl in functional tests from
tests/clientcmd/virtctl.go to tests/virtctl/cmd.go and update the
functional tests of virtctl to make use of the unexported helpers. This
prevents importing these helpers in other places where they should not
be used.

This also addresses a small bug in newVirtctlCommand, that was checking
the Value interface instead of the flag returned by flag.Lookup for nil.

Signed-off-by: Felix Matouschek <fmatouschek@redhat.com>