tests, libnet: Introduce new network-attachment-definition helpers
The helpers that have assisted with defining and creating the network
attachment definition (NAD) resource have been replaced by an improved
structured helpers.

E.g. instead of using a REST based client, a client-go one is used.

All callers who used the old helpers have been updated to use the new
ones. Some indirect usages have been kept to minimize the change.

The use of the `tuning` CNI plugin has been removed, as it was mainly
used to setup the `mtu` value for plugins that have not supported that
option. However, the current used CNI plugins do support it natively.
If a plugin does not support it, we should use it only if the test
validates it.

Signed-off-by: Edward Haas <edwardh@redhat.com>