guest-agent cmd utils: minor optimization
Currently, in order to check if all the commands are supported,
we check for each required command if that command is present and
enabled on the available command list. We are using two nested for
loop which result in O(n*m).

Creating an initial map of the available commands will result in
O(n+m).

Signed-off-by: fossedihelm <ffossemo@redhat.com>