heartbeat: fix possible shutdown race
The intent of the `heartBeatDone` channel was
to ensure the heartbeat is finished
before stopping the handler vm controller.
In particular, its intention was to prevent
a shutdown race:
- heartbeat tries to stop and do a last http request
- controller shuts down
If the controller shuts down first, then the main function
just exists, killing all remaining go routines right away.

Actually it is not doing its job since
`c.heartBeat.Run(c.heartBeatInterval, stopCh)` is
not a blocking function (it spawns a goroutine and
immediately returns the channel).
As result, the `heartBeat` is always immediately
closed after the start of the heartBeat goroutine.

To ensure that the heartbeat is completed before shutting
down the controller, we should block on the channel returned
by the heartbeat itself.

Signed-off-by: fossedihelm <ffossemo@redhat.com>