heartbeat: fix possible handler stuck
When the heartbeat is notified of a shutdown
by the stopchan, it marks the node unschedulable
before returning and thus end the entire process.

This labelling process is performed in a goroutine
which at the end signal the job done closing the
`done` channel.

This logic actually is not necessary, since it can
be a simple body function, without the goroutine.

In addition, there isn't a proper error handling.
In an error occurs during the goroutine execution
(i.e. node patch failure), nothing is returned
and the channel is never closed, which cause the
`<-labelNodeDone` to hangs and leave the virt-handler
stuck.

With this commit both problem are handler:
1. drop the unneeded goroutine
2. handle the error retrying it for a `DefaultBackoff`
(https://github.com/kubernetes/client-go/blob/37045084c2aa82927b0e5ffc752861430fd7e4ab/util/retry/util.go#L38)

Signed-off-by: fossedihelm <ffossemo@redhat.com>