vmi-ctrl, updateStatus: Report PodIfaceNames of secondary nets
Following PR #13018 the VMI controller reports the `PodInterfaceName`
field for the VMI's primary interface status.

Report the `PodInterfaceName` for secondary interfaces as well.
Doing so will enable propagating the naming scheme
(network names to pod interface names) from virt-controller to
virt-handler and virt-launcher and thus making this calculation
centralized instead of distributed in three places as we have today.

Signed-off-by: Orel Misan <omisan@redhat.com>