virt-launcher, net setup: Skip networks with unknown PodInterfaceName
There are scenarios on which the virt-launcher could not infer the
network naming scheme from the VMI.Status as it is not reported for some
or all interfaces:
1. An upgrade from an older version.
2. NIC Hotplug.

Filter the Spec interfaces and networks, and only operate on interfaces
and networks whose PodInterfaceName is known.

The VMI controller will eventually report the `PodInterfaceName`
for networks that were previously skipped, and they will eventually
be handled.

Signed-off-by: Orel Misan <omisan@redhat.com>