tests, net, sriov: Remove use GetRunningVMIDomainSpec
GetRunningVMIDomainSpec is a function that is used to access DomainXML
which is an intermediate technical detail that should be validated in
UT and does not belong in e2e functionality tests.
In addition, in order to reduce code dependencies, remove one use of
GetRunningVMIDomainSpec which comes from test.utils, a module which
we strive to remove all depedndecies of.
This test needed to use the function solely for the purpose of
obtaining the interface PCI address.
Instead of obtaining the PCI address assigned by libvirt, explictly
set it, and use it in the validation of the cloud-init data

Signed-off-by: Nir Dothan <ndothan@redhat.com>